[feat] OnboardingView에서 Avatar 등장 및 퇴장 효과음 추가
